+++ libcloud/storage/providers.py
@@ -52,6 +52,8 @@ DRIVERS = {
     ('libcloud.storage.drivers.ktucloud', 'KTUCloudStorageDriver'),
     Provider.AURORAOBJECTS:
     ('libcloud.storage.drivers.auroraobjects', 'AuroraObjectsStorageDriver'),
+    Provider.BACKBLAZE_B2:
+    ('libcloud.storage.drivers.backblaze_b2', 'BackblazeB2StorageDriver'),
 
     # Deprecated
     Provider.CLOUDFILES_US: